Move http actions to farmModule.
