+++ src/core/store/farmModule.js
@@ -4,6 +4,7 @@ import farmLog from '../../utils/farmLog';
 import defaultResources from './defaultResources';
 import { getRemoteLogs, sendRemoteLogs } from './http/sync';
 import createQuery from '../../utils/createQuery';
+import farm from './farmClient';
 
 // A function for updating an existing array of objects with an array of new
 // objects. An identifier is provided to determine if each new object is an
@@ -143,6 +144,50 @@ export default {
       return getRemoteLogs(context, payload)
         .then(() => sendRemoteLogs(context, payload))
         .then(() => { localStorage.setItem('syncDate', Math.floor(Date.now() / 1000)); });
+    },
+    updateFarmResources({ commit, dispatch }, response) {
+      const replaceResources = (res) => {
+        if (res.resources) {
+          commit('setFarmResources', res.resources);
+          dispatch('cacheFarmResources', res.resources);
+        }
+        return res;
+      };
+      if (response) {
+        return replaceResources(response);
+      }
+      return farm().info().then(replaceResources);
+    },
+    updateAreas({ commit }) {
+      return farm().area.get().then((res) => {
+        // If a successful response is received, delete and replace all areas
+        commit('deleteAllAreas');
+        const areas = res.list.map(({ tid, name, geofield }) => ({ tid, name, geofield }));
+        commit('addAreas', areas);
+      });
+    },
+    updateAssets({ commit }) {
+      return farm().asset.get().then((res) => {
+        // If a successful response is received, delete and replace all assets
+        commit('deleteAllAssets');
+        commit('addAssets', res.list);
+      });
+    },
+    updateUnits({ commit }) {
+      // Return units only.
+      return farm().term.get('farm_quantity_units').then((res) => {
+        commit('deleteAllUnits');
+        const units = res.list.map(({ tid, name }) => ({ tid, name }));
+        commit('addUnits', units);
+      });
+    },
+    updateCategories({ commit }) {
+      // Return categories only.
+      return farm().term.get('farm_log_categories').then((res) => {
+        commit('deleteAllCategories');
+        const cats = res.list.map(({ tid, name }) => ({ tid, name }));
+        commit('addCategories', cats);
+      });
     },
   },
 };
